Major changes in App.js
Deleted "Main.js" and imported all code into "App.js" due to how little the project is.

This resulted in bugs all across the code. All of them are fixed.

Added some functionality to the game. Now, you can guess letters and it displays if you were right or wrong.