Project Complete
Project completed.

It is a basic "hangman" game, which picks a random word from a large array. Then it displays this word, but the letters are hidden, it is up to you to guess them, but you only have 8 tries before you lose. Above the uknown word are 9 programming languages, which represent the "hangman". With each wrong guess a programming language gets lost, until you are finally left with Assembly, so you better guess correctly.

Final additions:.

Bugs, duplicates fixed,
Keyboard locks if won/lost,
Individual keyboard element locks if it is a wrong letter to prevent repeated or accident double clicks,
Styling of all elements, including the unknown word,
Added a "words" array,
Lots of conditional rendering added